agents: web, documents, memory and email namespaces run on the capability registry (PR 6 of tool-class retirement)
Ports the routed search trio, browser/screenshot, http/download, the
five document converters, thread memory, and email. Non-tool exports
(htmlToText, requestSignal, serpApiConfigured) move with their
implementations; ungatedCapabilityRun keeps the six context-only run
closures to one helper.
