+++ packages/agents/src/tools/thread-memory-tools.ts
@@ -2,441 +2,81 @@
  * Thread-memory tools — durable, per-conversation memory an agent manages
  * explicitly.
  *
- * These persist to the relational {@link ThreadMemory} store scoped to the
- * current chat thread (`context.threadId`). Unlike the in-session
- * `memory_list` / `memory_read` / `memory_write` tools (ephemeral, one agent
- * run) and the vector-backed `ltm_recall` / `ltm_remember` tools (fuzzy,
- * cross-session, embedding-gated), thread memories are deterministic, editable
- * rows that live and die with the thread — and each one can reference
- * resources of any kind (the assets it generates, a workflow it built, a
- * collection, an external URL) by a typed `{ type, id }` handle, so an agent
- * can record and reuse them across a creative project.
+ * These persist to the relational `ThreadMemory` store scoped to the current
+ * chat thread (`context.threadId`). Unlike the in-session `memory_list` /
+ * `memory_read` / `memory_write` tools (ephemeral, one agent run) and the
+ * vector-backed `ltm_recall` / `ltm_remember` tools (fuzzy, cross-session,
+ * embedding-gated), thread memories are deterministic, editable rows that live
+ * and die with the thread — and each one can reference resources of any kind
+ * (the assets it generates, a workflow it built, a collection, an external URL)
+ * by a typed `{ type, id }` handle, so an agent can record and reuse them
+ * across a creative project.
  *
- * Tools:
- *   - `thread_memory_save`   — write a memory, optionally referencing resources.
- *   - `thread_memory_list`   — list the thread's memories with resolved refs.
- *   - `thread_memory_update` — edit an existing memory by id.
- *   - `thread_memory_delete` — remove a memory by id.
+ * The four tools are @deprecated: they are ported to the `memory` capability
+ * module (`../capabilities/memory.ts`) and survive as thin subclasses so
+ * existing constructors keep working. The prompt renderer below is not a tool
+ * and stays here.
  */
 
-import type { ProcessingContext } from "@nodetool-ai/runtime";
-import { Asset, ThreadMemory } from "@nodetool-ai/models";
-import type { ThreadMemoryKind, ThreadMemoryResource } from "@nodetool-ai/models";
-import { Tool } from "./base-tool.js";
+import type { ThreadMemoryResource } from "@nodetool-ai/models";
+import { CapabilityTool, ungatedCapabilityRun } from "../capabilities/index.js";
+import {
+  threadMemoryDelete,
+  threadMemoryList,
+  threadMemorySave,
+  threadMemoryUpdate
+} from "../capabilities/memory.js";
+import type { Tool } from "./base-tool.js";
 
-const VALID_KINDS: ReadonlySet<string> = new Set([
-  "note",
-  "fact",
-  "preference",
-  "decision",
-  "resource"
-]);
-
-/** Known resource kinds — advisory; any string is accepted. */
-const KNOWN_RESOURCE_TYPES = [
-  "asset",
-  "workflow",
-  "collection",
-  "node",
-  "job",
-  "timeline",
-  "script",
-  "storyboard",
-  "image_document",
-  "thread",
-  "url",
-  "other"
-];
-
-const KIND_SCHEMA = {
-  type: "string" as const,
-  enum: ["note", "fact", "preference", "decision", "resource"],
-  description:
-    "Category of the memory. Use 'resource' when the point is the referenced " +
-    "resource(s), 'preference'/'decision'/'fact' for durable project context, " +
-    "else 'note'. Defaults to 'note'."
-};
-
-const RESOURCES_SCHEMA = {
-  type: "array" as const,
-  items: {
-    type: "object" as const,
-    properties: {
-      type: {
-        type: "string" as const,
-        description:
-          "Resource kind — one of: " +
-          KNOWN_RESOURCE_TYPES.join(", ") +
-          ". Any other value is allowed too."
-      },
-      id: {
-        type: "string" as const,
-        description:
-          "Identifier: asset id, workflow id, collection name, node type, a URL, etc."
-      },
-      uri: {
-        type: "string" as const,
-        description: "Optional canonical uri (asset://…, https://…)."
-      },
-      label: { type: "string" as const, description: "Optional human label." }
-    },
-    required: ["type", "id"]
-  },
-  description:
-    "Typed references to resources this memory is about — the assets you " +
-    "generated (type 'asset'), a workflow you built ('workflow'), a collection, " +
-    "a URL, etc. — so you can find and reuse them later. Asset references are " +
-    "validated and resolved to their asset:// uri; other kinds are stored as-is."
-};
-
-function coerceKind(value: unknown): ThreadMemoryKind {
-  if (typeof value !== "string") return "note";
-  const lower = value.toLowerCase().trim();
-  return (VALID_KINDS.has(lower) ? lower : "note") as ThreadMemoryKind;
-}
-
-/** Build the canonical `asset://<id>.<ext>` uri for an asset. */
-function assetUri(asset: Asset): string {
-  const ext = asset.fileExtension;
-  return ext ? `asset://${asset.id}.${ext}` : `asset://${asset.id}`;
+/**
+ * `thread_memory_save` — persist a memory to the current thread.
+ *
+ * @deprecated Ported to the `memory` capability module.
+ */
+export class ThreadMemorySaveTool extends CapabilityTool {
+  constructor() {
+    super(threadMemorySave.spec, threadMemorySave.impl, ungatedCapabilityRun);
+  }
 }
 
 /**
- * Normalize incoming resource refs from tool params. Asset refs are validated
- * against the user's library — unknown/foreign asset ids are dropped and
- * reported; every other kind is kept as-is (external URLs, resources the model
- * layer can't cheaply verify). `uri`/`label` for assets are backfilled.
+ * `thread_memory_list` — list the current thread's memories.
+ *
+ * @deprecated Ported to the `memory` capability module.
  */
-async function normalizeResources(
-  userId: string,
-  raw: unknown
-): Promise<{ resources: ThreadMemoryResource[]; dropped: ThreadMemoryResource[] }> {
-  if (!Array.isArray(raw)) return { resources: [], dropped: [] };
-  const resources: ThreadMemoryResource[] = [];
-  const dropped: ThreadMemoryResource[] = [];
-  for (const value of raw) {
-    if (!value || typeof value !== "object") continue;
-    const obj = value as Record<string, unknown>;
-    const type = typeof obj.type === "string" ? obj.type.trim() : "";
-    const id = typeof obj.id === "string" ? obj.id.trim() : "";
-    if (!type || !id) continue;
-    const ref: ThreadMemoryResource = {
-      type,
-      id,
-      ...(typeof obj.uri === "string" && obj.uri ? { uri: obj.uri } : {}),
-      ...(typeof obj.label === "string" && obj.label ? { label: obj.label } : {})
-    };
-    if (type === "asset") {
-      const asset = await Asset.find(userId, id);
-      if (!asset) {
-        dropped.push(ref);
-        continue;
-      }
-      ref.uri = assetUri(asset);
-      if (!ref.label) ref.label = asset.name;
-      ref.metadata = { content_type: asset.content_type };
-    }
-    resources.push(ref);
+export class ThreadMemoryListTool extends CapabilityTool {
+  constructor() {
+    super(threadMemoryList.spec, threadMemoryList.impl, ungatedCapabilityRun);
   }
-  return { resources, dropped };
 }
 
 /**
- * Resolve stored resource refs for display: re-check asset existence (dropping
- * assets the user no longer owns) and refresh their uri/label; pass every other
- * kind through unchanged.
+ * `thread_memory_update` — edit an existing memory.
+ *
+ * @deprecated Ported to the `memory` capability module.
  */
-async function resolveResources(
-  userId: string,
-  resources: ThreadMemoryResource[] | null | undefined
-): Promise<ThreadMemoryResource[]> {
-  if (!Array.isArray(resources) || resources.length === 0) return [];
-  const out: ThreadMemoryResource[] = [];
-  for (const ref of resources) {
-    if (!ref || typeof ref !== "object") continue;
-    if (ref.type === "asset") {
-      const asset = await Asset.find(userId, ref.id);
-      if (!asset) continue;
-      out.push({
-        type: "asset",
-        id: asset.id,
-        uri: assetUri(asset),
-        label: ref.label || asset.name,
-        metadata: { content_type: asset.content_type }
-      });
-    } else {
-      out.push(ref);
-    }
-  }
-  return out;
-}
-
-function requireThread(
-  context: ProcessingContext
-): { userId: string; threadId: string } | { error: string } {
-  const userId = context.userId;
-  if (!userId) return { error: "No user context; cannot access thread memory." };
-  const threadId = context.threadId;
-  if (!threadId) {
-    return {
-      error:
-        "No active thread; thread memory is only available inside a chat thread."
-    };
-  }
-  return { userId, threadId };
-}
-
-function droppedNote(dropped: ThreadMemoryResource[]): Record<string, unknown> {
-  return dropped.length > 0
-    ? {
-        dropped_resources: dropped,
-        note: "Some asset references were not found and were dropped."
-      }
-    : {};
-}
-
-/** `thread_memory_save` — persist a memory to the current thread. */
-export class ThreadMemorySaveTool extends Tool {
-  readonly name = "thread_memory_save";
-  readonly description =
-    "Save a durable memory to the current conversation. Use it to remember " +
-    "project facts, user preferences, decisions, and — crucially — the " +
-    "resources you produce or rely on (pass them in `resources`: the assets " +
-    "you generate, a workflow you built, a collection, a URL) so you can reuse " +
-    "them later. Memories persist across turns and are shown back to you at " +
-    "the start of each turn.";
-  readonly jsonSchema = {
-    type: "object" as const,
-    properties: {
-      content: {
-        type: "string" as const,
-        description:
-          "The memory itself — a self-contained note (e.g. 'The hero image " +
-          "uses a teal/orange palette the user approved')."
-      },
-      title: {
-        type: "string" as const,
-        description: "Optional short label shown when memories are listed."
-      },
-      kind: KIND_SCHEMA,
-      resources: RESOURCES_SCHEMA
-    },
-    required: ["content"] as string[]
-  };
-
-  async process(
-    context: ProcessingContext,
-    params: Record<string, unknown>
-  ): Promise<unknown> {
-    const scope = requireThread(context);
-    if ("error" in scope) return { success: false, error: scope.error };
-
-    const content = typeof params.content === "string" ? params.content.trim() : "";
-    if (!content) {
-      return { success: false, error: "content is required and must be a non-empty string" };
-    }
-    const { resources, dropped } = await normalizeResources(
-      scope.userId,
-      params.resources
+export class ThreadMemoryUpdateTool extends CapabilityTool {
+  constructor() {
+    super(
+      threadMemoryUpdate.spec,
+      threadMemoryUpdate.impl,
+      ungatedCapabilityRun
     );
-
-    try {
-      const memory = await ThreadMemory.create<ThreadMemory>({
-        user_id: scope.userId,
-        thread_id: scope.threadId,
-        kind: coerceKind(params.kind),
-        title: typeof params.title === "string" ? params.title.trim() : "",
-        content,
-        resources: resources.length > 0 ? resources : null
-      });
-      return {
-        success: true,
-        memory_id: memory.id,
-        kind: memory.kind,
-        resources,
-        ...droppedNote(dropped)
-      };
-    } catch (e) {
-      return { success: false, error: e instanceof Error ? e.message : String(e) };
-    }
-  }
-
-  userMessage(params: Record<string, unknown>): string {
-    const title = typeof params.title === "string" ? params.title : "";
-    return title ? `Remembering: ${title.slice(0, 60)}` : "Saving to memory";
   }
 }
 
-/** `thread_memory_list` — list the current thread's memories. */
-export class ThreadMemoryListTool extends Tool {
-  readonly name = "thread_memory_list";
-  readonly description =
-    "List the durable memories saved for the current conversation, newest " +
-    "first, each with its referenced resources resolved (asset references " +
-    "carry a live asset:// uri you can pass to view_image or reuse in " +
-    "generation tools).";
-  readonly jsonSchema = {
-    type: "object" as const,
-    properties: {
-      limit: {
-        type: "number" as const,
-        description: "Maximum memories to return (default 100, max 200)."
-      }
-    },
-    required: [] as string[]
-  };
-
-  async process(
-    context: ProcessingContext,
-    params: Record<string, unknown>
-  ): Promise<unknown> {
-    const scope = requireThread(context);
-    if ("error" in scope) return { success: false, error: scope.error };
-
-    const limitParam = Number(params.limit);
-    const limit =
-      Number.isFinite(limitParam) && limitParam > 0
-        ? Math.min(Math.floor(limitParam), 200)
-        : 100;
-
-    try {
-      const memories = await ThreadMemory.listByThread(
-        scope.userId,
-        scope.threadId,
-        limit
-      );
-      const items = [];
-      for (const memory of memories) {
-        items.push({
-          memory_id: memory.id,
-          kind: memory.kind,
-          title: memory.title,
-          content: memory.content,
-          created_at: memory.created_at,
-          updated_at: memory.updated_at,
-          resources: await resolveResources(scope.userId, memory.resources)
-        });
-      }
-      return { success: true, count: items.length, memories: items };
-    } catch (e) {
-      return { success: false, error: e instanceof Error ? e.message : String(e) };
-    }
-  }
-
-  userMessage(): string {
-    return "Recalling conversation memory";
-  }
-}
-
-/** `thread_memory_update` — edit an existing memory. */
-export class ThreadMemoryUpdateTool extends Tool {
-  readonly name = "thread_memory_update";
-  readonly description =
-    "Update a memory in the current conversation by id. Only the fields you " +
-    "pass are changed; pass `resources` to replace the referenced resources.";
-  readonly jsonSchema = {
-    type: "object" as const,
-    properties: {
-      memory_id: {
-        type: "string" as const,
-        description: "Id of the memory to update (from thread_memory_list)."
-      },
-      content: { type: "string" as const, description: "New content text." },
-      title: { type: "string" as const, description: "New title." },
-      kind: KIND_SCHEMA,
-      resources: RESOURCES_SCHEMA
-    },
-    required: ["memory_id"] as string[]
-  };
-
-  async process(
-    context: ProcessingContext,
-    params: Record<string, unknown>
-  ): Promise<unknown> {
-    const scope = requireThread(context);
-    if ("error" in scope) return { success: false, error: scope.error };
-
-    const memoryId = typeof params.memory_id === "string" ? params.memory_id : "";
-    if (!memoryId) {
-      return { success: false, error: "memory_id is required" };
-    }
-
-    try {
-      const memory = await ThreadMemory.find(scope.userId, memoryId);
-      if (!memory || memory.thread_id !== scope.threadId) {
-        return { success: false, error: `Memory not found: ${memoryId}` };
-      }
-
-      let dropped: ThreadMemoryResource[] = [];
-      if (typeof params.content === "string") memory.content = params.content.trim();
-      if (typeof params.title === "string") memory.title = params.title.trim();
-      if (params.kind !== undefined) memory.kind = coerceKind(params.kind);
-      if (params.resources !== undefined) {
-        const normalized = await normalizeResources(scope.userId, params.resources);
-        memory.resources = normalized.resources.length > 0 ? normalized.resources : null;
-        dropped = normalized.dropped;
-      }
-      await memory.save();
-
-      return {
-        success: true,
-        memory_id: memory.id,
-        resources: await resolveResources(scope.userId, memory.resources),
-        ...droppedNote(dropped)
-      };
-    } catch (e) {
-      return { success: false, error: e instanceof Error ? e.message : String(e) };
-    }
-  }
-
-  userMessage(): string {
-    return "Updating conversation memory";
-  }
-}
-
-/** `thread_memory_delete` — remove a memory. */
-export class ThreadMemoryDeleteTool extends Tool {
-  readonly name = "thread_memory_delete";
-  readonly description =
-    "Delete a memory from the current conversation by id when it's no longer " +
-    "relevant or was superseded.";
-  readonly jsonSchema = {
-    type: "object" as const,
-    properties: {
-      memory_id: {
-        type: "string" as const,
-        description: "Id of the memory to delete (from thread_memory_list)."
-      }
-    },
-    required: ["memory_id"] as string[]
-  };
-
-  async process(
-    context: ProcessingContext,
-    params: Record<string, unknown>
-  ): Promise<unknown> {
-    const scope = requireThread(context);
-    if ("error" in scope) return { success: false, error: scope.error };
-
-    const memoryId = typeof params.memory_id === "string" ? params.memory_id : "";
-    if (!memoryId) {
-      return { success: false, error: "memory_id is required" };
-    }
-
-    try {
-      const memory = await ThreadMemory.find(scope.userId, memoryId);
-      if (!memory || memory.thread_id !== scope.threadId) {
-        return { success: false, error: `Memory not found: ${memoryId}` };
-      }
-      await memory.delete();
-      return { success: true, memory_id: memoryId };
-    } catch (e) {
-      return { success: false, error: e instanceof Error ? e.message : String(e) };
-    }
-  }
-
-  userMessage(): string {
-    return "Forgetting a memory";
+/**
+ * `thread_memory_delete` — remove a memory.
+ *
+ * @deprecated Ported to the `memory` capability module.
+ */
+export class ThreadMemoryDeleteTool extends CapabilityTool {
+  constructor() {
+    super(
+      threadMemoryDelete.spec,
+      threadMemoryDelete.impl,
+      ungatedCapabilityRun
+    );
   }
 }
